Fix gcc 3 warning: comparison is always true due to limited range of data type

+++ hw/ide/internal.h
@@ -486,7 +486,7 @@ static inline IDEState *idebus_active_if(IDEBus *bus)
 
 static inline IDEState *bmdma_active_if(BMDMAState *bmdma)
 {
-    assert(bmdma->unit != -1);
+    assert(bmdma->unit != (uint8_t)-1);
     return bmdma->bus->ifs + bmdma->unit;
 }
